# openapi.yaml
openapi: 3.1.0
info:
  title: Affilite Mix API
  version: "1"
  description: "Multi-tenant affiliate marketing platform API. Routes marked x-internal: true are not part of the public consumer contract."
  contact:
    name: Affilite Mix
  license:
    name: Proprietary
servers:
  - url: "https://compareai.site"
    description: Production (compareai.site)
  - url: "https://staging.compareai.site"
    description: Staging
paths:
...
  /api/admin/content:
    get:
      summary: GET /api/admin/content (site) — auth=admin
      operationId: getApiAdminContent
      tags:
        - Admin
      x-auth: admin
      x-adminRequired: true
      x-tenantScope: site
      x-rateLimit: true
      x-csrf: true
      x-requestSchema: ContentInput
      x-responseSchema: Content
      x-notes: null
      security:
        - bearerAuth: []
      x-internal: true
      responses:
        200:
          description: Successful response (Content)
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
          content:
            application/json:
              schema:
                type: object
        401:
          description: Unauthorized
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        403:
          description: Forbidden — insufficient role or cross-tenant access
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        429:
          description: Rate limited
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
    post:
      summary: POST /api/admin/content (site) — auth=admin
      operationId: postApiAdminContent
      tags:
        - Admin
      x-auth: admin
      x-adminRequired: true
      x-tenantScope: site
      x-rateLimit: true
      x-csrf: true
      x-requestSchema: ContentInput
      x-responseSchema: Content
      x-notes: null
      security:
        - bearerAuth: []
      x-internal: true
      responses:
        200:
          description: Successful response (Content)
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
          content:
            application/json:
              schema:
                type: object
        401:
          description: Unauthorized
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        403:
          description: Forbidden — insufficient role or cross-tenant access
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        429:
          description: Rate limited
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
    patch:
      summary: PATCH /api/admin/content (site) — auth=admin
      operationId: patchApiAdminContent
      tags:
        - Admin
      x-auth: admin
      x-adminRequired: true
      x-tenantScope: site
      x-rateLimit: true
      x-csrf: true
      x-requestSchema: ContentInput
      x-responseSchema: Content
      x-notes: null
      security:
        - bearerAuth: []
      x-internal: true
      responses:
        200:
          description: Successful response (Content)
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
          content:
            application/json:
              schema:
                type: object
        401:
          description: Unauthorized
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        403:
          description: Forbidden — insufficient role or cross-tenant access
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        429:
          description: Rate limited
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
    delete:
      summary: DELETE /api/admin/content (site) — auth=admin
      operationId: deleteApiAdminContent
      tags:
        - Admin
      x-auth: admin
      x-adminRequired: true
      x-tenantScope: site
      x-rateLimit: true
      x-csrf: true
      x-requestSchema: ContentInput
      x-responseSchema: Content
      x-notes: null
      security:
        - bearerAuth: []
      x-internal: true
      responses:
        200:
          description: Successful response (Content)
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
          content:
            application/json:
              schema:
                type: object
        401:
          description: Unauthorized
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        403:
          description: Forbidden — insufficient role or cross-tenant access
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        429:
          description: Rate limited
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
...
components:
...
  headers:
    API-Version:
      description: Numeric major version of the served API contract.
      schema:
        type: string
        pattern: "^[1-9][0-9]*$"
        example: "1"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...